requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
clarifai-grpc==9.7.4
python-telegram-bot==13.15
pytz==2023.3
six==1.16.0
tornado==6.1
tzdata==2023.3
tzlocal==5.0.1
emoji==1.7.0
dnspython==2.4.2 
pymongo==4.5.0